+++ index.html
@@ -234,6 +234,8 @@
       </section>
     </template>
   <script src="js/functions.js"></script>
-  <script src="js/main.js"></script>
+  <script type="module" src="js/main.js"></script>
+  <script type="module" src="js/data.js"></script>
+  <script type="module" src="js/util.js"></script>
 </body>
 </html>
